Delete the dead wasm pipeline the v0 retirement orphaned along with the second Perceus, the sketch emitters, the unused wasm-encoder dependency, and five tracked binaries
