+++ .gitignore
@@ -17,3 +17,8 @@ docs-site/
 # The tree+toolchain fingerprint `make verify-trust` records so `make receipt`
 # can fold in a verification of the identical tree (proofs/lib/stamp.sh).
 proofs/.verified-fingerprint
+
+# Build artifacts must not be tracked (#930)
+*.cwasm
+*.rmeta
+spec/**/*.wasm
