+++ Cargo.toml
@@ -35,7 +35,6 @@ serde = { version = "1", features = ["derive"] }
 serde_json = "1"
 semver = "1"
 toml = "1.1"
-wasm-encoder = "0.254"
 lasso = { version = "0.7", features = ["multi-threaded"] }
 lsp-server = "0.10"
 lsp-types = "0.97"
