+++ crates/almide-codegen/Cargo.toml
@@ -11,7 +11,6 @@ almide-lang = { path = "../almide-lang" }
 almide-ir = { path = "../almide-ir" }
 almide-egg-lab = { path = "../almide-egg-lab" }
 egg = "0.11"
-wasm-encoder = "0.253"
 wasmparser = "0.254"
 toml = "1.1"
 serde_json = "1"
