+++ crates/almide-codegen/src/pass.rs
@@ -39,7 +39,6 @@ pub struct ScopeContext {
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub enum Target {
     Rust,
-    Wasm,
     Wgsl,
 }
