+++ crates/almide-dialect/src/lib.rs
@@ -14,7 +14,6 @@ pub mod ops;
 pub mod lower;
 pub mod verify;
 pub mod dump;
-pub mod emit_rust;
 
 use almide_base::intern::Sym;
